Change 7: Classes 'SimBox' and 'Sim' planned out
The Project object's command loop now enters a sub-loop if you select a
particular Region object to examine (and process with simulations).
Also, after discussing optimisation and scenario differences with
Romesh, I have created stubs for a SimBox and SimBox-contained Sim
class.
Functionality considerations are ongoing.
Note: Romesh has provided proper working default parameters for
makeworkbook.py, allowing files to be called something other than
'example'. These are stored in a new file named preloaded.py, which will
store default constants to be shared across multiple modules.
This change is not the focus of the OOP experiment, so be wary of
bugs...