Change 10: Default Sim object first steps
Regions are now able to create new SimBox objects and contain them.
Unlike with Regions, which can be made empty, the creation of a SimBox
forces a default Sim object to be constructed inside.
With data (D.data) and metadata (D.G) stored in the Region object, these
are passed down to the Sim so as to run makedatapars.py.
The resulting parsdata (D.P) is stored in the Sim object. Further
updates will store M and F as well.
Additionally, the Portfolio sub-loop examineregion(currentregion) method
now allows you to 'check' data and metadata with appropriate commands.