Change 14: SimBox inheritance
Optimisation container SimBoxOpt is now a derived class from SimOpt.
Portfolio.py will still ask you which of the two you want to create in
the region.
SimBoxOpt functionality is limited at the moment, but the portfolio will
not allow you to optimise a SimBox object (checked at Region level).
There is still the problem that optimisation requires a normal run to
generate D['S']. This must be solved before running a simulation can be
seen as either a normal run or an optimisation process.