Change 21: Optimising multiple regions
Portfolio GPA method now iterates through loaded regions, creates a
SimBox for each, creates a default SimBudget, then runs an optimisation
cycle. Multiplot is then run for each SimBox.
Note: It technically runs two optimisation cycles, but only because the
optimised SimBudget needs to be processed to get plot data. This can be
coded a better way later, so as not to run unnecessary cycles just to
get plot data.