Change 24: Attempt at objective gradient
My current idea for GPA involving numerous regions is to optimise each
and then work out how rapidly a diagnostic 'objective' value changes
with regard to variations in regional budget.
This is currently and tentatively called an objective gradient.
The plan is that money will be moved into regions where the gradient is
steepest and pulled out of regions where the gradient is most shallow,
resulting in a net optimisation.
This is only a first attempt at calculation though and more thought is
required.