Change 28: Testing objective curve optimisation
New file geoprioritisation.py is thus far a proof of principle.
It applies monotonic PCHIP fitting to two arrays of budget totals and
corresponding objective values. The sum of these splines is then
successfully minimised according to a sum budget total constraint.
The next step is to store the PCHIP splines in SimBoxes, generate them
for regions, and then optimise from the portfolio level.