Change 33: BALLSD GPA now acts on portfolio simboxes
Code has been tidied up so that the BALLSD GPA method no longer works
specifically on a hard coded example.
The gpaoptimisefixedtotal can now be called (without checks) on a list
of SimBoxOpts with BOCs.
Alternatively, the Portfolio class can call its geoprioanalysis method,
pick the first SimBoxOpt with a BOC in every stored region (or create a
new SimBoxOpt if necessary), and then run gpaoptimisefixedtotal.