Change 36: BOC migration to Region level
For reasons of data persistence and clarity, BOC data and associated
methods are now stored on the Region level, rather than SimBox level.
If a Region does not have a BOC, the developBOC method will create a
temporary SimBoxOpt and SimBudget to calculate one. They will then be
deleted.
GPA optimisation is currently commented out in the Portfolio class. It
only generates BOCs. Monotonicity restrictions are also disabled for the
moment.
Lastly, an attempt at parallelising BOC generation with the
multiprocessing module does not work in the Spyder IDE. This is a known
issue online, so commented code stubs are kept in this change, just in
case. They will be removed in the next change.