Change 38: GPA optimisation reintroduced
File geoprioritisation.py has been recoded to utilise Region-stored BOCs
as opposed to the old SimBoxOpt-stored BOCs.
Given that BOC calculation is technically separate from optimisation, no
Sims need to be created during GPA objective minimisation.
However, this will likely be induced in the next change so as to check
hypothesised objectives and visualise the effects of GPA.