Change 40: Beginning SimBudget/SimBudget2 fusion
In the process of working separately on overrides and optimisation,
Romesh and I have produced two versions of the SimBudget class.
Mine hooks into Cliff's optimisation routine but works with alloc, while
Romesh's builds model parameters from a full time-dependent budget
array.
This is a first attempt to combine both, although the makemodelpars is
currently having issues with the 'program' class and is commented out.
The budget variable is thus currently unused.
Note: The viewoptimresults function has been postponed as we can
probably generate a better plotting routine later.