Change 42: Revert to specialoptload for GPA purposes
With the Vancouver meeting and potential for GPA demonstration
approaching, optimised SimBudget creation has been reverted to use of
the 'specialoptload' method.
It uses Cliff's 'optimize' function as a black box to generate model
parameters for the new SimBudget.
However, it should be easy to switch back to Romesh's 'makemodelpars'
method once program CCOCs are fleshed out for the new OOP Optima.