Change 47: Solved model inconsistency bug
A major bug was previously fixed relating to model.py incorrectly
modifying input D.M values in a non-local manner. This was a typical
deepcopy issue and the augmented model parameters had an effect on
comparing OOP with legacy models.
However, plots of the two still differed.
This change finally solves that bug, by identifying that
['tmpbestdata'][-1]['R'] in optimize.py does not actually store the
results of the best optimisation, only the latest.
In short, this becomes a huge problem during uncertainty analysis.
Because of this, optimize.py now returns the direct result of a BALLSD
optimisation to SimBudget instead.
Plots now demonstrate consistency between default and barely-optimised
Simulations, indicating (finally) valid results.