Change 48: Backup before functionality layer shifts
Optimisation functionality is currently at the Sim level, while
calibration functionality is at the Region level. However, Romesh and I
are aiming to shift this into SimBoxOpt and a prospective SimBoxCal
class, respectively.
This change marks the last copy of Optima that 'runs' GPA before this
cleanup is attempted.
The validity of the process is in question though, with several parallel
issues to be handled.
At the moment, even legacy calibrated files are initiating uncertainty
analysis, presumably due to the new calibration implementation, which
makes BOC calculation unpredictable.