Change 54: GPA review and objective value calculations
Portfolio now stores a list relevant to GPA runs, specifically a list of
lists with references to SimBoxOpt objects used in the analyses.
This has allowed the development of Portfolio command 'review' to check
the results of GPAs.
There is no DataIO currently for this temporary storage of GPA results,
but it does make rechecking plots simpler.
Also, SimBudget now contains code relevant to determining the objective
value of a simulation without requiring optimisation. A lot of this code
has been duplicated from optimize.py.
Note: It is currently hard-coded to run with the default 'optimisation
options'. This will need to be changed later. Also, perhaps this
objective calculation can be applied in the core Sim class rather than
the derived SimBudget.