Change 57: Force Sim makemodelpars option
Class SimBudget is in the process of having an overloaded
initialisation, so that it runs initially as a Sim with a budgetless
model.
This will allow it to generate a D.S structure that can be used in
getcurrentbudget, as Robyn's develop branch changes require.
The process is not yet complete though and is extremely tricky to do
properly.