Change 61: Objective defaults
In order to make optimisation control somewhat easier, default objective
specifications have been moved to defaults.py.
This means that the objective calculations used in both optimize.py and
simbudget.py are consistent.
Note that normalisations work strangely, being 1 when there is only one
objective aim, and being far greater than 1 if there are multiple aims.
This may not be too big of an issue. It just means the outcome values
are somewhat odd.