Low-Funding and Multi-Objective Fixes
Iyanoosh and Azfar have highlighted that optimisation fails for a
certain project where a vital program is initially funded at $0.01.
This change implements an optimize.py fix to redistribute the original
allocation within the optimisation context, so as to avoid this
situation.
Multiple objectives are now also fixed so all targets are always
achieved in minimizemoney.py.
Allocation multiplications are now done in 1.2 steps rather than
doublings.