Popsizes Scaled
District popsizes have been scaled in both data and calibrations.
However, it is not clear by what factors other values should be scaled,
such as prevalence. For instance, if national total prevalence is 0.4%,
but total prevalence is 8% in a particular district, what should a
national MSM prevalence of 20% be scaled to? Are popsize values required
for scaling (e.g. via PLHIV formulations) even when they are not
available in Project.data?
This exercise is left to the developer.