Results fix and geospatial improvements
Tired of never being able to getresults() attributed to a parset because
its resultsref is a unicode, not a str?
Type checking for unicode now implemented in getresults().
Geospatial sheet construction now allows a year to be provided from the
tvec of the last calibration stored (e.g. 2000-2030) rather than just
data years (e.g. 2000-2015).
Pop and prev ratios for district project construction use the national
project's calibration as a denominator, as opposed to the ratio sum.
While this means that ratios will no longer add to 1 unless calibration
is matched very carefully, it does allow for districts to be excluded
(i.e. modelling 1/3rd of a country).