Improved BOC plotting
The geogui thus far only plotted BOCs, as well as points on the BOC that
would correspond to the pre-GA and post-GA budgets.
Now it is possible to plot extra points on the BOC plots, which is
currently used to pinpoint pre-GA pre-opt and post-GA post-opt
coordinates.
This should give an indication if the re-optimisation maxtime value was
sufficient during GA.