BOC Extrapolation Hotfix
The final point in a BOC calculation is generally chosen to be 10 times
the default district budget, i.e. a value considered large enough that
funding would not be pushed there.
However, during Jamaican analysis, this upper bound on some of the 13
districts was 10% of the national budget, i.e. easily allocated during a
random re-distribution.
This hot fix thus adds an extra point to the BOC whenever interpolation
is called for; specifically at (1e15,BOC.y[-1]).
This flatlines everything after 10 times the district budget and does
affect the BOC's curvature approaching the final point. However, these
minor shortcomings are preferable to everything being distributed to one
district due to unbounded cubic extrapolation.