Fix budget year and budget values not being arrays (#1885)
The FE expects both the budget year(s) and budget values to be arrays, not single numbers which the BE sometimes gives. This promotes the values to arrays if they are not None before passing to the FE. They stay as plain numbers in the BE until the FE clicks save. Commit list:
* Budget not being an array in the FE causes problems so try to promotetoarray before passing to FE

* Budget not being an array in the FE causes problems debug 1

* Budget not being an array in the FE causes problems so try to promote budget to array

* Revert "Budget not being an array in the FE causes problems debug 1"

* Clean up

* Don't promote to array if None

* Don't promote to array if None - cleaner